@@ -16,14 +16,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.2",
-    "@angular/cdk": "^12.0.2",
-    "@angular/common": "~12.0.2",
-    "@angular/core": "~12.0.2",
-    "@angular/forms": "~12.0.2",
-    "@angular/platform-browser": "~12.0.2",
-    "@angular/platform-browser-dynamic": "~12.0.2",
-    "@angular/router": "~12.0.2",
+    "@angular/animations": "^12.1.1",
+    "@angular/cdk": "^12.1.1",
+    "@angular/common": "~12.1.1",
+    "@angular/core": "~12.1.1",
+    "@angular/forms": "~12.1.1",
+    "@angular/platform-browser": "~12.1.1",
+    "@angular/platform-browser-dynamic": "~12.1.1",
+    "@angular/router": "~12.1.1",
     "@capacitor-community/firebase-analytics": "^1.0.0",
     "@capacitor-community/firebase-crashlytics": "^1.0.0",
     "@capacitor/app": "^1.0.0",
@@ -61,11 +61,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.2",
-    "@angular/cli": "~12.0.2",
-    "@angular/compiler": "~12.0.2",
-    "@angular/compiler-cli": "~12.0.2",
-    "@angular/language-service": "~12.0.2",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
+    "@angular/language-service": "~12.1.1",
     "@capacitor/cli": "^3.0.0",
     "@ionic/angular-toolkit": "^2.3.0",
     "@ngxs/devtools-plugin": "^3.7.1",
