@@ -1,11 +1,11 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {TeamMatchesEntry} from '../../../../../core/api/models/team-matches-entry';
+import {Component, EventEmitter, Input, OnInit, Output, ViewEncapsulation} from '@angular/core';
 import {Player} from '../../../../../core/api/models/player';
 
 @Component({
     selector: 'beping-team-match-player-item',
     templateUrl: './team-match-player-item.component.html',
-    styleUrls: ['./team-match-player-item.component.scss']
+    styleUrls: ['./team-match-player-item.component.scss'],
+    encapsulation: ViewEncapsulation.None
 })
 export class TeamMatchPlayerItemComponent implements OnInit {
 
